+++ Plugin.Notifications/Platforms/Android/CrossNotifications.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+
+namespace Plugin.Notifications
+{
+    public static partial class CrossNotifications
+    {
+        static CrossNotifications()
+        {
+            Current = new NotificationsImpl();
+        }
+    }
+}