+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGE LOG
 
 ## 5.0.0
+* [feature] .NET Standard 2.0 (and latest android support libs)
 * [feature] Notification events!
-* [feature] Control the notification icon used in Android ()
+* [feature] Control the notification icon used in Android
+* [feature] Control launch intent flags on Android
+* [fix][ios] Metadata wasn't being sent
 * [fix][android] cancelling a single notification would not remove it from the schedule database
 
 ## 4.0.2